move from utils to native
