+++ tests/ignite/distributed/comp_models/test_native.py
@@ -9,7 +9,45 @@ from ignite.distributed.comp_models import has_native_dist_support
 if not has_native_dist_support:
     pytest.skip("Skip if no native dist support", allow_module_level=True)
 else:
-    from ignite.distributed.comp_models.native import _NativeDistModel
+    from ignite.distributed.comp_models.native import _expand_hostlist, _NativeDistModel
+
+
+# tests from https://github.com/LLNL/py-hostlist/blob/master/hostlist/unittest_hostlist.py
+@pytest.mark.parametrize(
+    "hostlist, expected",
+    [
+        ("localhost", "localhost"),
+        ("quartz[4-8]", "quartz4,quartz5,quartz6,quartz7,quartz8"),
+        (
+            "node[18-19,1-16,21-22]",
+            "node1,node2,node3,node4,node5,"
+            "node6,node7,node8,node9,node10,"
+            "node11,node12,node13,node14,node15,"
+            "node16,node18,node19,node21,node22",
+        ),
+        (
+            "node[4-8,12,16-20,22,24-26]",
+            "node4,node5,node6,node7,node8,"
+            "node12,node16,node17,node18,"
+            "node19,node20,node22,node24,"
+            "node25,node26",
+        ),
+        ("machine2-[02-4]vm1", "machine2-02vm1,machine2-03vm1,machine2-04vm1"),
+        (
+            "machine2-[02-3]vm1, machine4-[0003-5].vml2",
+            "machine2-02vm1,machine2-03vm1," "machine4-0003.vml2," "machine4-0004.vml2," "machine4-0005.vml2",
+        ),
+        ("machine2-[009-11]vm1", "machine2-009vm1,machine2-010vm1,machine2-011vm1"),
+        ("node[1,2,3]", "node1,node2,node3"),
+    ],
+)
+def test_expand_hostlist(hostlist, expected):
+    assert _expand_hostlist(hostlist) == expected.split(",")
+
+
+def test_expand_hostlist_unvalid():
+    with pytest.raises(ValueError, match=r"hostlist unvalid"):
+        _expand_hostlist("unvalid[]")
 
 
 @pytest.mark.distributed
